Work around padding-bottom not working on Firefox
When the output on the board is long enough to enable vertical srollbar,
the last line of output appears to stick to the bottom edge of the
board. This issue occurs because the bottom padding of 1em is not
honoured by Firefox when the scrollbar appears. See
https://stackoverflow.com/q/13471910 for a detailed discussion on this
issue.

This change works around this issue by retaining the margin for the last
child inside the board. The last child is usually a paragraph with a
vertical margin of 1em. Now the output appears good on Firefox with the
last line of text neatly separated from the bottom edge of the board
by 1 em. However, on Chrome the last line of text is separated from the
bottom edge of the board by 2em (1em padding and 1em margin). This is
still better than the issue of no vertical space at all after the last
line on Firefox.

To summarize, this change makes a choice between having no vertical
space after the last line on the board on Firefox and extra vertical
space on Chrome and the latter is chosen as a compromise.