Keep CSS selectors for board elements unnested
Since most of the CSS code is meant for styling the output elements to
be displayed inside the output board, style such elements with simple
CSS selectors (not nested selectors).

When the same elements need to be styled differently outside the output
board, those elements can be put inside a different container div
element and nested selectors can be used to style such elements.