Make it easy to type Markdown within ,center block
The ,center editing command currently inserts the following HTML code:

    <div class="center">

    </div>

The cursor is placed on the blank line between the open and close
tags, so when the user starts typing on this blank line, the text is
not interpreted as Markdown. See
https://spec.commonmark.org/0.29/#html-blocks for an explanation of
this behaviour in the CommonMark specification.

As per the CommonMark specification, we need to insert a blank line
after opening the above HTML tag to resume Markdown again. Therefore,
in this change, the ,center command inserts the following HTML code:

    <div class="center">

    </div>

The cursor is now left at the second blank line after the open tag.
This allows any text inserted there to be interpreted as Markdown.