Fix disappearing output on Firefox on Android
The following line of code in `replaceCommand` removes an input
command from the input text field:

    input.value = s.substr(0, textPos) + text1 + text2 +
                  s.substr(cursorPos)

On most web browsers (e.g., Firefox on macOS, Chrome on macOS and
Android), this change of input field by JavaScript does not trigger an
input event. However, with Firefox 87 on Android 10, this line of code
triggers an additional input event. This behaviour was found only with
Firefox on Android. As a result, the `handleKey` function gets invoked
again which schedules another delayed render. Thus after a command
like `,help` has finished displaying the help message, the delayed
render kicks in and clears the board. This causes the output of a
command like `,help` to disappear soon after it is displayed.

Note that the above code removes the command being processed from the
text input field.

This change fixes this issue by clearing the delayed render timer
immediately after the input text field is changed to remove the input
command.